database re-creation. saving the number of products and hours

+++ Controllers/LKController.cs
@@ -48,13 +48,13 @@ namespace BilliardClub.Controllers
         {
             var user = await _userManager.GetUserAsync(User);
             var orders = await _context.Orders
+                .Where(x => x.user.Id == user.Id)
                 .Include(x => x.user)
-                .Include(x => x.poolTables)
-                .Include(x => x.restaurantMenu)
-                .Where(x => x.user.Id == user.Id).ToListAsync();
+                .Include(x => x.poolTables).ThenInclude(x => x.poolTable)
+                .Include(x => x.foodItems).ThenInclude(x => x.foodItem)
+                .ToListAsync();
             return PartialView(orders);
         }
-
 
         [HttpGet]
         public async Task<IActionResult> _AccountSettings()
@@ -80,10 +80,30 @@ namespace BilliardClub.Controllers
         }
 
         [HttpPost]
+        public void UpdateQuantity(int cartItemId, uint quantity)
+        {
+            if (_cart.CartItems.Exists(x => x.id == cartItemId))
+            {
+                if (quantity <= 0)
+                    quantity = 1;
+
+                if (quantity > 15)
+                    quantity = 1;
+
+                var cartItem = _context.CartItems.First(x => x.cartId == _cart.cartId && x.id == cartItemId);
+                cartItem.quantity = quantity;
+                _context.SaveChanges();
+            }
+        }
+
+        [HttpPost]
         public void DeleteTableInCart(int id)
         {
-            var table = _context.PoolTables.FirstOrDefault(x => x.id == id);
-            _cart.DeleteTableInCart(table);
+            if (_cart.CartItems.Exists(x => x.PoolTable != null && x.PoolTable.id == id))
+            {
+                var table = _context.PoolTables.FirstOrDefault(x => x.id == id);
+                _cart.DeleteTableInCart(table);
+            }
         }
 
         [HttpPost]
@@ -102,10 +122,33 @@ namespace BilliardClub.Controllers
         public async Task ConfirmOrder(double cheque)
         {
             var order = new Order() {orderDate = DateTime.Now, cheque = cheque, user = await _userManager.GetUserAsync(User)};
-            var tables = _cart.CartItems.Where(x => x.PoolTable != null).Select(x => x.PoolTable).ToList();
-            var products = _cart.CartItems.Where(x => x.FoodItem != null).Select(x => x.FoodItem).ToList();
-            order.poolTables.AddRange(tables);
-            order.restaurantMenu.AddRange(products);
+            var tablesInCart = _cart.CartItems.Where(x => x.PoolTable != null).ToList();
+            var productsInCart = _cart.CartItems.Where(x => x.FoodItem != null).ToList();
+
+            if (tablesInCart.Count != 0)
+            {
+                foreach (var tableInCart in tablesInCart)
+                {
+                    order.poolTables.Add(new OrderPoolTable()
+                    {
+                        poolTable = tableInCart.PoolTable,
+                        quantity = tableInCart.quantity
+                    });
+                }
+            }
+
+            if (productsInCart.Count != 0)
+            {
+                foreach (var productInCart in productsInCart)
+                {
+                    order.foodItems.Add(new OrderFoodItem()
+                    {
+                        foodItem = productInCart.FoodItem,
+                        quantity = productInCart.quantity
+                    });
+                }
+            }
+
             _context.Orders.Add(order);
             _cart.DeleteAllItemsInCart();
             await _context.SaveChangesAsync();
